+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and as of version 1.0.0, follows semantic versioning.
 
 ## [Unreleased]
   * Fixed dependency versions to speed up `pip install`
+  * Allow Dashi to pull from more than one cache source
+  * Update Dashi to gsiqcetl V1.21
 
 ## [230330-1838] - 2023-03-30
   * Reverted to rnaseqqc2 V2 cache due to column naming bug